Split things up a bit and made the UI more clear #50
