+++ app/workouts/templates/set/panel.html
@@ -1,0 +1,4 @@
+<h4 class="list-group-item-header">
+  <%- weight %>
+</h4>
+<p class="list-group-item-text"><%- reps %> Reps</p>
